Add v1 router
